feat: decimal energy fields, import wizard fixes and UI enhancements                                                                                                                                                                    - Change total_energy_consumption_kwh_per_year and annual_energy_consumption_kwh     from PositiveIntegerField to DecimalField(max_digits=12, decimal_places=3)                                                                                       across AC, chiller, ventilation, lighting, HWS and lift models                                                                                                 - Migration 0030: AlterField for all 6 energy consumption fields                                                                                                 - Update clean() methods in lighting, lift and HWS forms to accept Decimal                                                                                       - Update all JS auto-calc helpers to use .toFixed(3) for energy fields                                                                                             and .toFixed(2) for total_power_input_kw (ventilation)                                                                                                         - Update import backend _calc_*_energy helpers to return Decimal(3dp)   - Fix structural component import: derive input_unit from assembly dimension     instead of Excel unit column (fixes 'percent' validation error)   - Fix lift import: upsert instead of rejecting duplicate (second import updates)   - Fix renewable_energy_percent validation: allow 0 (was rejecting values < 0.01)   - Add baseline_efficiency field to boiler and water heater UI templates   - Add system tab row counts to import map-fields step   - Add collapsible system tab sections with horizontal scroll to import preview step   - Fix structural preview: show meaningful columns (component, EPD name, country, qty)

+++ pages/models/building_operation/air_conditioning.py
@@ -98,7 +98,9 @@ class CoolingSystemAirConditioner(models.Model):
     baseline_leakage_factor_percent = models.PositiveIntegerField(
         default=2, verbose_name=_("Baseline Leakage Factor (%)")
     )
-    total_energy_consumption_kwh_per_year = models.PositiveIntegerField(
+    total_energy_consumption_kwh_per_year = models.DecimalField(
+        max_digits=12,
+        decimal_places=3,
         null=True,
         blank=True,
         verbose_name=_(
